pyproject.toml
[project]
name = "siteship-api"
version = "0.1.0"
description = ""
authors = [
    {name = "samrat", email = "samratshakya5@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
    "playwright (>=1.44.0,<2.0.0)",
    "supabase (>=2.16.0,<3.0.0)",
    "python-telegram-bot (>=22.2,<23.0)",
]
